Fixed uses of uninitialized variables in emailer and privmsg
privmsg.php was trying to set a number of template parameters based on
variables that were nowhere set. I've just deleted those lines, since
the template does not try to use these parameters anyway.

emailer.php now initializes the 'to', 'cc' and 'bcc' as needed later on
in the code.